Add bzlmod-prebuilt-modules/local-module-extensions
This adds a new BzlMod experiment that tries to use module extensions
to create the repositories for the project's dependencies, instead
of using local_path_override() in the project's top-level.

Ideally, this could be done programatically with a Starlak function
which could read a manifest file listing all the external repositories
and their location, but for now just hard-code the paths in the
module extension function.

This does *not* work though because native.local_repository() isn't
supported yet in extension functions.